Add support for Dockerfile directives
Because of issues in docker-file-parser:
  https://github.com/joyent/node-docker-file-parser/issues/8
we have to first keep track of dockerfile directives, and fold
them into the parsing performed by the external module.

We also have to keep track of the escape directive, so that we can
ensure we parse comments in the middle of commands correctly.

This work leads the way to being able to specify livepush behaviour
using dockerfile directives.

Change-type: patch
Signed-off-by: Cameron Diver <cameron@balena.io>