Add support for livecmd
This is a new method of allowing Dockerfiles to specificy a command
that will be run when the container is started with livepush. We
generate a new Dockerfile with this command in mind, and replace
the internal representation with this but also return it so the
build process can use that.

Change-type: minor
Signed-off-by: Cameron Diver <cameron@balena.io>